03-Deep-Learning-Projects/01-Computer-Vision-Projects/03-Transfer-Learning-ResNet/requirements.txt
tensorflow>=2.13.0
numpy>=1.24.0
pillow>=10.0.0
matplotlib>=3.7.0
pytest>=7.4.0
pytest-cov>=4.1.0 